693 binary number with alternating bits
I was thinking it's easy to cheat at first because integer range's
alternating bit value is limited. So using the calc() method to export
a list of alternating bits values. When the main method invoked, only
need to do a binary search.
But the fastest solution is using XOR to calc the result like this:

n = (n ^ (n>>1));
return (n & (n+1)) == 0;

That's really brilliant.